Update Tally.java
fix ec pairing bug
Modified to generate multiple voter messages.
default is to create a message of 1000 voters, and the number of voters can be set as an argument. ex) examples.generators.Tally 10000
In the real tally protocols, EC_V, EC_W must be decrypted through java_test/DecryptMessage.java to find out the msg, and then the calculation must be performed.